refactor: create allDecksCounts
GSoC 2025: Review Reminders

- Created allDecksCounts scheduler method to abstract out the logic of counting all cards that are due across the entire app, as this logic is used in both WidgetStatus (also updated in this PR) and NotificationService (coming in a later PR).
